+++ decider-bcl2fastq/src/test/java/ca/on/oicr/pde/deciders/Bcl2fastqDeciderTest.java
@@ -14,6 +14,8 @@ import ca.on.oicr.pde.client.SeqwareClient;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedSet;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Lists;
 import com.google.common.collect.Table;
 import com.google.common.io.Files;
 import java.io.File;
@@ -22,6 +24,7 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.EnumMap;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.SortedMap;
@@ -38,6 +41,7 @@ import org.mockito.Mockito;
 import static org.mockito.Mockito.when;
 import org.powermock.reflect.Whitebox;
 import static org.testng.Assert.assertEquals;
+import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
@@ -109,6 +113,8 @@ public class Bcl2fastqDeciderTest {
                 .sequencerRunPlatformModel("HiSeq")
                 .createdDate(expectedDate)
                 .rootSampleName("TEST_0001")
+                .sampleName("TEST_0001_001")
+                .iusTag("AAAAAAAA")
                 .sampleAttributes(sp1Attrs)
                 .skip(false)
                 .provenanceId("1")
@@ -135,6 +141,8 @@ public class Bcl2fastqDeciderTest {
                 .sequencerRunPlatformModel("NextSeq")
                 .createdDate(expectedDate)
                 .rootSampleName("TEST_0001")
+                .sampleName("TEST_0001_001")
+                .iusTag("AAAAAAAA")
                 .sampleAttributes(sp2Attrs)
                 .skip(false)
                 .provenanceId("2")
@@ -146,24 +154,41 @@ public class Bcl2fastqDeciderTest {
         when(lpp.getLaneProvenance()).thenReturn(Arrays.asList(lp1, lp2));
     }
 
+    @AfterMethod
+    public void destroySeqware() {
+        Whitebox.<Table>getInternalState(MetadataInMemory.class, "STORE").clear();
+    }
+
     @Test
     public void all() throws IOException {
         //run on all lanes
         assertEquals(bcl2fastqDecider.run().size(), 2);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 2);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 2);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 4);
 
         //all lanes have been scheduled - no new bcl2fastqWorkflow runs expected
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 4);
 
         //rerun everything
         bcl2fastqDecider.setIgnorePreviousAnalysisMode(true);
         assertEquals(bcl2fastqDecider.run().size(), 2);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 2);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 2);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 8);
 
         //all should be blocked again
         bcl2fastqDecider.setIgnorePreviousAnalysisMode(false);
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 8);
     }
 
@@ -171,10 +196,16 @@ public class Bcl2fastqDeciderTest {
     public void beforeDateFilter() {
         bcl2fastqDecider.setBeforeDateFilter(expectedDate.minusDays(1));
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 0);
 
         bcl2fastqDecider.setBeforeDateFilter(expectedDate.plusDays(1));
         assertEquals(bcl2fastqDecider.run().size(), 2);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 2);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 2);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 4);
     }
 
@@ -182,10 +213,16 @@ public class Bcl2fastqDeciderTest {
     public void afterDateFilter() {
         bcl2fastqDecider.setAfterDateFilter(expectedDate.plusDays(1));
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 0);
 
         bcl2fastqDecider.setAfterDateFilter(expectedDate.minusDays(1));
         assertEquals(bcl2fastqDecider.run().size(), 2);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 2);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 2);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 4);
     }
 
@@ -193,10 +230,16 @@ public class Bcl2fastqDeciderTest {
     public void includeInstrumentFilter() {
         bcl2fastqDecider.setIncludeInstrumentNameFilter(ImmutableSet.of("does not exist"));
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 0);
 
         bcl2fastqDecider.setIncludeInstrumentNameFilter(ImmutableSet.of("h001"));
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0001"));
@@ -207,6 +250,9 @@ public class Bcl2fastqDeciderTest {
     public void excludeInstrumentFilter() {
         bcl2fastqDecider.setExcludeInstrumentNameFilter(ImmutableSet.of("h001"));
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0002"));
@@ -219,10 +265,16 @@ public class Bcl2fastqDeciderTest {
         filters.put(FileProvenanceFilter.sequencer_run, ImmutableSet.of("does not exist"));
         bcl2fastqDecider.setIncludeFilters(filters);
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 0);
 
         filters.put(FileProvenanceFilter.sequencer_run, ImmutableSet.of("RUN_0001"));
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0001"));
@@ -235,6 +287,9 @@ public class Bcl2fastqDeciderTest {
         filters.put(FileProvenanceFilter.sequencer_run, ImmutableSet.of("RUN_0001"));
         bcl2fastqDecider.setExcludeFilters(filters);
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0002"));
@@ -247,10 +302,16 @@ public class Bcl2fastqDeciderTest {
         filters.put(FileProvenanceFilter.lane, ImmutableSet.of("does not exist"));
         bcl2fastqDecider.setIncludeFilters(filters);
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 0);
 
         filters.put(FileProvenanceFilter.lane, ImmutableSet.of("RUN_0001_lane_1"));
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0001"));
@@ -263,6 +324,9 @@ public class Bcl2fastqDeciderTest {
         filters.put(FileProvenanceFilter.lane, ImmutableSet.of("RUN_0001_lane_1"));
         bcl2fastqDecider.setExcludeFilters(filters);
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0002"));
@@ -275,10 +339,16 @@ public class Bcl2fastqDeciderTest {
         filters.put(FileProvenanceFilter.sequencer_run_platform_model, ImmutableSet.of("does not exist"));
         bcl2fastqDecider.setIncludeFilters(filters);
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 0);
 
         filters.put(FileProvenanceFilter.sequencer_run_platform_model, ImmutableSet.of("HiSeq"));
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0001"));
@@ -291,6 +361,9 @@ public class Bcl2fastqDeciderTest {
         filters.put(FileProvenanceFilter.sequencer_run_platform_model, ImmutableSet.of("HiSeq"));
         bcl2fastqDecider.setExcludeFilters(filters);
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0002"));
@@ -303,10 +376,16 @@ public class Bcl2fastqDeciderTest {
         filters.put(FileProvenanceFilter.study, ImmutableSet.of("does not exist"));
         bcl2fastqDecider.setIncludeFilters(filters);
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 0);
 
         filters.put(FileProvenanceFilter.study, ImmutableSet.of("TEST_STUDY_1"));
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0001"));
@@ -319,6 +398,9 @@ public class Bcl2fastqDeciderTest {
         filters.put(FileProvenanceFilter.study, ImmutableSet.of("TEST_STUDY_1"));
         bcl2fastqDecider.setExcludeFilters(filters);
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
         for (FileProvenance fp : getFpsForCurrentWorkflow()) {
             assertEquals(fp.getSequencerRunNames(), ImmutableSet.of("RUN_0002"));
@@ -335,22 +417,34 @@ public class Bcl2fastqDeciderTest {
         bcl2fastqDecider.setDoCreateIusLimsKeys(true);
         bcl2fastqDecider.setDoScheduleWorkflowRuns(true);
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
 
         bcl2fastqDecider.setIsDryRunMode(false);
         bcl2fastqDecider.setDoCreateIusLimsKeys(true);
         bcl2fastqDecider.setDoScheduleWorkflowRuns(false);
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
 
         bcl2fastqDecider.setIsDryRunMode(false);
         bcl2fastqDecider.setDoCreateIusLimsKeys(true);
         bcl2fastqDecider.setDoScheduleWorkflowRuns(true);
         assertEquals(bcl2fastqDecider.run().size(), 0);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 0);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
 
         bcl2fastqDecider.setIsDryRunMode(false);
         bcl2fastqDecider.setDoCreateIusLimsKeys(true);
         bcl2fastqDecider.setDoScheduleWorkflowRuns(true);
         bcl2fastqDecider.setIgnorePreviousLimsKeysMode(true);
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
 
         Collection<FileProvenance> fps = getFpsForCurrentWorkflow();
         assertEquals(fps.size(), 2);
@@ -369,6 +463,9 @@ public class Bcl2fastqDeciderTest {
 
         //sp2 has two group ids, can not be run
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 1);
         assertEquals(getFpsForCurrentWorkflow().size(), 2);
 
         //correct the group ids
@@ -376,7 +473,46 @@ public class Bcl2fastqDeciderTest {
         sp2Attrs.put(Lims.GROUP_ID.getAttributeTitle(), ImmutableSortedSet.of("group2"));
 
         assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
         assertEquals(getFpsForCurrentWorkflow().size(), 4);
+    }
+
+    @Test
+    public void duplicateBarcodeTest() {
+        SampleProvenanceImpl.SampleProvenanceImplBuilder sp = SampleProvenanceImpl.builder()
+                .sequencerRunName("RUN_0002")
+                .studyTitle("TEST_STUDY_2")
+                .laneNumber("1")
+                .sequencerRunPlatformModel("NextSeq")
+                .createdDate(expectedDate)
+                .rootSampleName("TEST_0003")
+                .sampleName("TEST_0003_001")
+                .iusTag("AAAAAAAA")
+                .sampleAttributes(sp2Attrs)
+                .skip(false)
+                .provenanceId("9")
+                .version("1")
+                .lastModified(expectedDate);
+
+        List<SampleProvenance> currentList = Lists.newArrayList(spp.getSampleProvenance());
+        when(spp.getSampleProvenance()).thenReturn(Lists.newArrayList(Iterables.concat(currentList, Arrays.asList(sp.build()))));
+
+        assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 1);
+        assertEquals(getFpsForCurrentWorkflow().size(), 2);
+
+        //correct the barcode and schedule
+        sp.iusTag("TTTTTTTT");
+        when(spp.getSampleProvenance()).thenReturn(Lists.newArrayList(Iterables.concat(currentList, Arrays.asList(sp.build()))));
+        assertEquals(bcl2fastqDecider.run().size(), 1);
+        assertEquals(bcl2fastqDecider.getScheduledWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getValidWorkflowRuns().size(), 1);
+        assertEquals(bcl2fastqDecider.getInvalidWorkflowRuns().size(), 0);
+        assertEquals(getFpsForCurrentWorkflow().size(), 5);
     }
 
     private Collection<FileProvenance> getFpsForCurrentWorkflow() {