+++ decider-bcl2fastq/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1 - 2017-03-15
+- Update to pipedev 2.1 (provenance-client 1.1)
+- [GP-1022](https://jira.oicr.on.ca/browse/GP-1022) - Validate barcodes before scheduling
+- [GP-1040](https://jira.oicr.on.ca/browse/GP-1040) - Reduce memory usage
+- [GP-1082](https://jira.oicr.on.ca/browse/GP-1082) - Modify decider to remain in error state if there is an error during workflow run scheduling
 ## 1.0 - 2017-02-02
 - Initial java implementation of the bcl2fastq/CASAVA decider
 - [GP-867](https://jira.oicr.on.ca/browse/GP-867) - Add support to create IUS-LimsKey, a SeqWare object that is used to link LIMS data from Pinery and SeqWare to SeqWare workflow runs